Tested (and tweaked) DownloadTitle. There is a call to download iconoclasts right before the main window pops up. The download happens in the background so it doesn't wait to download the entire game before starting. There is no progress indicator, and the downloaded files are of size 0 until they are complete.
I'd like to find a way to display and update file progress, and write the downloaded files to disk in chunks, instead of all at once, so that you can see the progress on disk and so that you can resume files.

Verify is not implemented yet.
Creating CNMT XML is not implemented.
NSP packing is not implemented.
Choosing a title to download via UI is not implemented.