Fixed a major bug when downloading large files. Waiting for the content meant the whole file would download before the content even returned. Waiting for the headers only led to errors if I didn't want to read the content byte by byte (like getting back really small results, or just headers).
The big fixes were twofold.

One was making it optional whether to wait for content (default) or not (only for DownloadFile calls).

The second was that I was using an HttpClient in a using statement, as show in online examples, but that meant it was disposed before the long-term async download operation was completed, leading to an error.

Online sources say that async and using don't work well together. I can see why.

I changed the HttpClient to be a singleton, so it is never disposed. This saves some time and memory from not having to make it all the time, and never disposing it allows HttpCompletionOption to work properly.

Updating and tagging to 0.2.0. Downloading all NCAs in a game works now.

I'll tag to 0.3.0 after I get progress bars and NSP repacking working, probably.