Fixed a bug in the last update that prevented resuming of files.
This was actually several bugs.

One, I thought File.OpenWrite would open for appending if the file existed, but it seems it was opening at the start. I had to File.Open() and specific FileMode.Append.

Second, I still had a call to make a HEAD request in DownloadFile, which was part of an experiment but was supposed to be changed back to GET.

Third, I had some issues reading headers, but I realized that you can get headers more easily via built-in properties, rather than accessing them by string. Content-Range wasn't being found properly, and now it is, and Content-Length is automatically parsed into a long so I don't have to do it myself.

Finally, I updated DownloadTask to allow specifying a starting point. Progress when a file resumes will now start where it is supposed to. Updated the progress callbacks. Also fixed an extra percent in there.