v0.2.5
Added a text field at the top to filter by game name.

Added saving and loading library metadata. The favorite checkbox was already working automatically through Data Binding, but now it saves to the file so it is permanent. Once library scanning is up and running you'll be able to generate the metadata for your library (file location, size, favorite) without manual editing, which is what I'm doing right now.

Changed the setting for library file to not have an extension. I may change it to automatically choose a format based on the extension, with no extension using a default format, but for now it is always xml and gets an xml extension.

Note: Library XML contains no title keys or names. It contains IDs to tie each entry to the appropriate title. KEYS are still loaded from titlekeys.txt.