v0.2.7 Implemented update title keys. Downloads new title keys file according to the URL in settings and displays new titles in a message box. It also sets the state to "new". I still have to write code that sets it to not new after some time passes, I think, probably on load.
I do not know if it properly updates the datagrid in real time or if you have to wait until the app restarts. I haven't tested that.

I noticed that the images of the new games didn't show up. I think that's because I still only have image preload. I don't have anything that triggers on viewing an image or clicking on a game and downloads the image if it is missing. So, they'll be blank unless I trigger a full preload.

I need to show stuff in a better format than a plain message box, because it goes off the screen if there are too many new titles.

There's a weird bug where if I let the datagrid autosize itself to the window everything slows to a crawl. If I restrict the height to a constant number, it runs fine. Of course then the height is either too big (runs out of the window) or too small. I've set it to 600, which is small, but it still runs off the end of the window for some reason. Maybe my margins are set weird.