Oops, I didn't actually get a new cert, I just repacked the old cert. Here is the new one.
Also committed changes without testing them, since I was assured it would work. The new asynchronous observable collection works now. I don't reallly understand it, but the collection was hanging every time something was added to it. I don't know if this is worse than the problem it was supposed to solve, which was an exception when trying to add to the collection from a different thread. It was solved in the same way, by adding all items at once from a temp collection rather than adding to the observable collection one at a time.

Also updated firmware and deviceId settings.