In trying to fix the bugs in observable collection by creating a new one and swapping (instead of adding to the existing one, because apparently that isn't allowed depending on the thread trying to do it), I created a new bug where I was bypassing the adding of titles to the title index. Thus, the titles weren't found in the index and re-added later when DLC was added, since it didn't know which title to add the DLC to and made a stub.
Since I just got rid of observable collection and replaced it with a List, I rewrote the code to just call AddGame.

Also fixed a bug where the library xml was not getting truncated, so if the library shrunk for any reason you'd end up with extra text at the end of the XML file and fuck up the parsing.

Also some refactoring I missed.