v.0.8.0
Lots of UI fixes and improvements.

You can now click on a title icon to open up its eshop page.

I reorganized and fleshed out the Menus. Many are still not implemented, but they are clearly labeled and they show a text box explaining themselves when clicked.

Revamped the filtering. All checkboxes are for enabling the view of that type. Check the demos box to show demos, and so on. Favorites is an exception, as it filters only by favorites. Otherwise you wouldn't be able to show only favorites within the other categories, it would show, for example, favorites AND owned games, instead of favorites that are owned.

It is now possible to edit the collection state (owned, not owned, etc) as a drop down directly in the grid. It is also possible to change the file path.

I caused and fixed a bug that prevented properties from being updated from the grid to the data model. I don't know what even caused it, as it was working most of the time and then suddenly it wasn't. The fix was simple but I don't know what it was needed in the first place. Unintended consequence of some other change I assume.

Found a way to do "one click editing". DataGrid normally requires you to click a cell, then click the control within the cell to modify it, then "deselect" the entire row before the cell can be editing and the value pushed out to the data model. I made it so that editing is started and ended by mouseover, so that by the time you click the editing starts immediately, and it ends and is sent out once you leave. One unfortunate consequence of this is that you must keep the mouse over a text string while you type a new value, but (assuming I don't find a fix) it is worth the convenience.

The two new classes related to enums are there as a system to bind enums to comboboxes, which isn't built-in to wpf for some reason. See the various Enum classes and the UI xaml for examples on how to bind the enum and modify the selected value and the text show in the combo box.

Fixed NSP creation so that illegal characters are removed from the name of the NSP before creating the file name. Otherwise there is an error when creating the NSP file if there are slashes or colons and some other characters. They are removed, not replaced.

Removed the Downloaded SwitchCollectionState. It is redundant, same meaning as Owned, and I don't know why I made it. Maybe if I remember why and it was a good reason I'll add it back in.

Moved Favorite column to the start of the row.

Reduced some font sizes.

Fixed exception when checkbox was modified before the datagrid was initialized.

Updated library xml to output the game name, which can make hand editing easier (if you want to).