v0.8.2
Images now load if the row is clicked on but the image was never preloaded. Blank images should never show up unless there is no image.

As of now, a blank image object is always used, and checking a blank image triggers a remote check. That means if an image does NOT exist at all, it will be looked for every time.

Ideally I'd fix it so that null signifies the file was never loaded remotely and blank image signifies that a remote image load was attempted but there was no image, and I would only try a lookup if it was null. That is a deeper change so it will be done later, or never.