v0.8.5
I attempted to write code connecting to the eshop, only to find that the current method of connecting no longer works... Not even Shogun works. All requests are rejected.

So I went with my backup plan, which involves writing a function to calculate the NSPs size without actually downloading the content. It adds up the Content-Length headers for each NCA, plus the CNMT size, the CNMT xml, the cert and the ticket. It all goes pretty fast, and I save all of the numbers in the XML library file so it doesn't need to be calculated again.

This doesn't take into account updates or DLC - just the version 0 base game that will be downloaded when you download base game only.

If you have a file, the size is still calculated using the actual on disk file size.