v0.8.6
Added progress bars for downloads! The coolest part about it is that I was able to use data binding to attach the DownloadTask to both the progress bar and the text so that they were auto updated by the value of progress changing, and I didn't actually have to use the download progress callback or write any code at all to update the progress value.

I'm not sure if the download speed is working, it is a bit wonky and WIP.

The progress bars work great, and the download window pops up on download start. Progress bars appear for EVERY download, but the window only appears for title downloads, and bars disappear very quickly for small files.

This included updating DownloadTask to implement IPropertyChanged so it could notify changes for Binding.

I also added a Downloads menu, with a button for showing and  hiding the downloads window.

Updated the title of the window so it isn't called MainWindow anymore.

Refactored gameCollection to library.