v0.8.7
Double clicking a progress bar cancels the download. I haven't properly tested it, but it should also stop the nsp from being repacked, and result in a null NSP as if it was never set to be repacked.

I may make a better version of cancelling where an exception is thrown and caught later.