v0.9.0
Added filtering for New titles. To change titles to not new, either download them (marks them as owned) or manually change to not owned, owned or on switch. Titles are auto marked as New as soon as they are added via title key update and stay that way until changed.

Changed downloads to keep downloads on completion and added a clear button. Not tested!