v1.0.1
Loads of UI features, useability updates and bug fixes.

Added a Readme, finally. Not complete but lots of stuff.

Fixed reference to download progress event.

Made some library elements nullable, then changes library metadata loading so only non null elements are copied into the current library.

Let you set a temp path for downloading NCAs rather than putting them in the nsp directory. I have NOT made it so that the NCA folder is copied from the temp folder to the nsp folder when not repacking - it stays in the temp folder.

Added a right click context menu that lets you download, toggle favorites, delete a title (a duplicate or bad title, for example), remove ownership of a title, and more things to come probably.

Changed the switch collection back to an observablecollection for the millionth time, since I need it to make the DataGrid update when an item is removed.

BoxArt is now properly saved and permanently replaces the Icon, if it exists. It also notifies an update to Icon in addition to BoxArtUrl.

Disallow closing or resizing the progress window. I may re-allow resizing later, I dunno.

Added button in Downloads menu to cancel a bulk download.

Fixed some bugs so now bulk downloading of favorites definitely works. Not sure yet about updates or DLC.

Added grid sorting in code behind. Sorting now defaults to Name (ascending), but I could save the most recent sort and restore it in a future update.

Fixed an error that crashed the app in a rare instance where the details element of the grid was NULL, but the "visibilitychanged" action went off.

Some refactoring and removing dead references.