Fixed bug where downloading DLC or updates tried to get images out of the control NCA... but the control NCA was null... and then we tried to delete a null file... and instead it deleted all of my project files. Yes, this happened 3 times before I found and fixed the bug. Only try to get the control nca if it exists.
Added support for sending the current downloading title to the DownloadFile function so it is available later to the progress window.

Removed inappropriate CertificateDeniedException (boy was that confusing) where a CnmtMissingException was more appropriate.

For some reason the size parsed by the CNMT content entry is COMPLETELY wrong for DLC. I would estimate file sizes using nca.Size, get a very small number (~100k), then proceed to download it and find it is 100 MB. So, when estimating DLC size use the old way, where I call GetContentLength.