Added LatestVersion to SCI and LibraryMetadata so it can be saved. Now, the most recent update will be loaded up with the library without querying nintendo's servers.
Derp. Fixed name of BoxArtUrl (xml name should be BoxArt to match LibraryMetadata). That's why boxart kept disappearing.

Changed title and library loading to NOT update versions. Versions are merely left at default (for new titles) and loaded from XML (for library metadata). That also means that loading titles and metadata is no longer async, since there's no web call. I could rewrite the file loading to be async, but it happens so fast that what would be the point, really? The UI would response a fraction of a second earlier, but the operation would be slower (due to overhead) and the library would be empty. I guess you could watch the items get added to the UI as they are loaded, which would look cool, but it would happen nearly instantly.

Added XAML event handlers for Loading and Closing window (Loading is new, moved library loading to it, and Closing is a replacement for the old override).

Added the VersionsConverter to the versions combobox.

Fixed some bugs where UI methods were called without the dispatcher, and updated other dispatcher calls to use InvokeOrExecute. This includes the new ShowError method (and updated ShowMessage method) that uses the dispatcher to call a message box.