Fixed an irritating bug that's been bothering me since since I made DLOption databound. The problem is that the visibilitychanged event didn't seem to actually happen early enough to update the current details row, so DLOption would be blank... then it would whatever the LAST row you clicked on's default is.
I now have the even PreviewLeftMouseButton or some nonsense, and it fires the second I click on the row and before the details open. I change the DLOption to match the type (game or dlc) and it updates appropriately.

Also tweaked the way the version is gotten, since now dlc latestversion should be preloaded.