Updated control.nacp parsing. Mostly finished.
I also went ahead and added code to update the name and publisher of the title any time the control NCA is decrypted, using the values in the file.

Added an "Added" field to SwitchCollectionItem, which tracks when the NSP was added to the library.

Fixed bugs in creating NSPs, related to recent changes to how NSPs and CNMTs work.

Added some useful extensions for hashing and converting to and from hex.

Fixed bug where metadata was saved to a file with no extension if manually saved with Library -> Save Library Metadata.