v1.0.7
Support for remembering your filtering and grid sorting preferences between sessions.

Fixed some small bugs in parsing control.nacp.

I spent a lot of time and wrote a lot of code searching for a major threading bug that caused seemingly random crashes when trying to update a title's size or icon. I tried making the delete functions smarter, fucked with the timing of threads and lots of other things.

Eventually I discovered it was due to the details visibility changed event firing multiple times, and thus calling the updatesize function twice really fast. It would try to access the same files at the same time, and one would try to delete something that had just been deleted.

I suspect this bug has been there for some time, and that random crash I had when loading icons or sizes was this same bug.

I fixed it by 1) only running the event handling code when the details were visible, so it only ran once and 2) fixing the real problem, which is allowing two threads to work on the same folder at the same time. I ended up using semaphores that are unique to a title + base directory + version (which is essentially what defines a title download or folder). The entire method must take the semaphor before it can work on the directory, but if the semaphore exists for another title download, it can go on ahead without a problem.

Moved some utility functions to a new FileUtils class. Also added a bunch of crap hoping to fix the problems with conflicts in opening files. They had nothing to do with the (threading) problem but they're fine I guess.

There were also some changes to jobs because for a sec I thought it was a good idea for the job to dispose of the stream, but I thought better of it. Streams should still be disposed by whoever created the job.