v1.0.7a
Two big fixes. One is that somehow the code for importing a client cert disappeared. The cert was banned today and I found that out. Fortunately the Crypto code was still there so restoring the event handler was easy.

Second, because I was using "using" with the semaphore for locking, the semaphore was disposed. The next time that same lock was acquired, it tried to get a disposed lock. That has been fixed.

I also put a "lock" block around the semaphore dictionary, so a thread can't swoop in and steal a lock while you're checking for and trying to acquire that lock.

Added a client password setting (not used yet).