Finally implemented getting the dauth token. It works!
I have changed the eshop cert to just be a regular-ass cert, because that's what it requires. It doesn't use shopN.pem. The current public cert is banned (it says so in the response to the challenge), so I used my personal cert for the EShopCert and it worked.

I'm currently hardcoding the password to the eshop cert as "switch".