Implemented eshop api login and data reading, then my cert was banned, so I implemented eshop web page parsing instead.
They're both still there. If the setting "LoginCertPath" points to a certificate, it will try to access the api. If the setting is missing, or the file is missing, or the login fails due to banning, it will swap over to using html parsing to get the data.

I also added a few more bits of data to the metadata.

Added MaxConcurrentDownloads. If you try to download more titles than this (defaults to 4), the new downloads will silently queue in the background until a slot is available. This is 4 "download commands", so a single title and all of its updates count as a single synchronous job.

Added a little more data to the title display.

Fixed a bug where New Preload was being changed to Preload on restart.

Made Unlockable count as new, so if a new key is loaded and you check "New", unlockable titles will show up.

Stop loading invalid title IDs (Armello had an ID of 16 0s).

Changed NumPlayers to be a number instead of a description, to better match the API. That was while it still worked, but I left it that way.

Better error handling and logging.